finelog: pin k8s deploy image to its digest at apply time
`finelog deploy` now resolves the configured image tag (e.g. `:latest`) to its
immutable `@sha256:` digest before rendering the k8s manifests, so the
Deployment pins an exact image. The config keeps the readable `:latest`, but a
redeploy lands precisely what the tag points to now and the running spec records
the digest — no node-cache staleness.

- extract `resolve_image_digest` (+ `_extract_digest`) from `_gcp` into a shared
  `finelog.deploy.image` module; reuse it in the k8s backend (`k8s_up`,
  `k8s_restart`) and keep the existing GCP call sites
- deployment template keeps `imagePullPolicy: IfNotPresent` (cache-safe once the
  image is a digest, matching iris task pods; only the iris controller uses
  Always because it does not resolve digests)
- resolution falls back to the tag with a warning if docker/registry is
  unavailable
