+++ lib/finelog/deploy/k8s/02-deployment.yaml.tmpl
@@ -33,6 +33,10 @@ spec:
         fsGroup: 1000
       containers:
         - name: finelog
+          # `finelog deploy` resolves the configured tag (e.g. `:latest`) to an
+          # immutable `@sha256:` digest before rendering, so this is a pinned
+          # reference and IfNotPresent is cache-safe — no node-cache staleness,
+          # and the spec records exactly what is running.
           image: {{ image }}
           imagePullPolicy: IfNotPresent
           # SYS_PTRACE lets memray/py-spy attach to the running server for
